fix: add global: global to access global functions in sandbox-js
