+++ backend/src/routes/client/code.route.js
@@ -1,0 +1,7 @@
+const express = require('express');
+const router = express.Router();
+const controller = require('../../controllers/client/code.controller');
+
+router.post('/execute',  controller.execute)
+
+module.exports = router
