+++ backend/src/routes/client/index.route.js
@@ -3,6 +3,7 @@ const oauthRoutes = require('./oauth.route')
 const tokenRoutes = require('./token.route');
 const postRoutes = require('./post.route');
 const profileRoutes = require('./profile.route');
+const codeRoutes = require('./code.route')
 
 module.exports = app => {
   const version = '/api/v1';
@@ -12,4 +13,5 @@ module.exports = app => {
   app.use(version + '/token', tokenRoutes);
   app.use(version + '/forum', postRoutes);
   app.use(version + '/profile', profileRoutes);
+  app.use(version + '/code', codeRoutes)
 };
